Add files via upload
Pseudocode outlining my approach to Project One. The target of the assignment is to create a program that holds data on courses and has a menu to execute different commands. The main commands that will be in the menu are load data, print courses in alphanumeric order, and print specific course information.